+++ R/gg_pointrange.R
@@ -25,7 +25,6 @@
 #'     ymin = lower,
 #'     ymax = upper,
 #'     position = position_dodge(width = 0.2),
-#'     size = 0.2,
 #'     x_title = "Treatment",
 #'     y_title = "Response",
 #'   )